Refactor todo and get all todo data
