+++ components/Todo.module.css
@@ -1,0 +1,33 @@
+.bttn {
+  padding: 15px 40px;
+  background: #6491C9;
+  color: white;
+  text-transform: uppercase;
+  font-weight: 400;
+  font-size: 1.2em;
+  letter-spacing: 2px;
+  transition: 0.5s;
+  margin-right: -1.3em;
+  margin-left: -1.3em;
+  clip-path: polygon(90% 0, 100% 50%, 90% 100%, 0 100%, 10% 50%, 0 0);
+
+  @media screen and (max-width: 760px) {
+    padding: 8px 15px;
+    font-size: 0.5em;
+    letter-spacing: 0.5px;
+    margin-right: -1em;
+    margin-left: -1.2em;
+  }
+}
+
+.bttn:hover {
+  cursor: pointer;
+  background: #1D1E1C;
+  color: #E5E6E7;
+  letter-spacing: 4px;
+  clip-path: polygon(100% 0, 90% 50%, 100% 100%, 10% 100%, 0% 50%, 10% 0);
+
+  @media screen and (max-width: 760px) {
+    letter-spacing: 2px;
+  }
+}
